// backend/src/controllers/types/authTypes.ts
export type GitHubTokenResponse = {
    access_token?: string;
    token_type?: string;
    scope?: string;
    error?: string;
    error_description?: string;
};

export type GitHubUserResponse = {
    login: string;
    id: number;
    node_id: string;
    avatar_url: string;
    gravatar_id: string;
    url: string;
    html_url: string;
    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscriptions_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_events_url: string;
    type: string;
    user_view_type: string;
    site_admin: boolean;
    name: string | null;
    company: string | null;
    blog: string;
    location: string | null;
    email: string | null;
    hireable: boolean | null;
    bio: string | null;
    twitter_username: string | null;
    notification_email: string | null;
    public_repos: number;
    public_gists: number;
    followers: number;
    following: number;
    created_at: string;
    updated_at: string;
};

export type GitHubEmailResponse = {
    email: string;
    primary: boolean;
    verified: boolean;
    visibility: 'public' | 'private' | null;
};